adjust sign in error code notification
